Remove debian/install and debian/postinst
debian/install is necessary when you have multiple binary packages from
a single source package. Since we don't have that here, we don't need
to modify installation locations.

debian/postinst is for when you need additional site-specific
configuration per package. In this case, it was running the
setup.py script again. This is bad because it installs a second
copy of the package into /usr/local, which shouldn't happen. When the
debian package is removed, these files will remain, preventing full
removal. Packages should install directly to /usr and should not
create any files in /usr/local.