Remove desktop file
We don't need a desktop file for Repoman in Pop OS because we will be
using Pop_Shop to open it directly. If we think it makes sense to have
an entry for Software Sources in the main menu, we can keep it, but
this makes little sense for our use case.