Suppress error dialog for denied by policy
This happens when a user cancels the authentication request. We probably
don't need to throw and error dialog for this action, since the user clearly
canceled the request. We do, however, log this to the console in case the user
is an advanced user, or in case of support where it's unknown by the user
why the action didn't take.