fix: Remove double-click handlers
The double click handlers cause issues on the Flatpak page because
double-clicking is an affirmative action, while delete is destructive.

We remove the double-click on the Extra Sources page to keep consistency with the
Flatpak page. Otherwise, the user may become frustrated when the double click
works on the Extra Sources page, but not on the Flatpak page, since they are
mostly identical widgets.