fix(source-errors): display dialog only on initial run
If there are errors with source files, only display an error dialog on the first time the application opens. Otherwise, we might spam the user with error dialogs. With this change, the dialog will spawn if there are errors when the application first opens, and a WARNING log output will be printed to the console every time the application subsequently tries to load the sources.