fix(list): set add button to sensitive regardless of how adding a source turns out
The add button gets set to insensitive when a dialog opens, and is normally set sensitive when the source lists change. However, if a dialog (like the add dialog) were opened and did not finish successfully, the button would not be set sensitive until the lists change again. This prevented users from adding a new source if they add one that errors.